Small Bugfix on core hpp.

+++ sdk/clean.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-make clean
+if [ -f Makefile ] ;then
+  make clean
+  rm Makefile
+fi
+
 rm CMakeCache.txt
 rm -rf CMakeFiles
 rm cmake_install.cmake
 rm install_manifest.txt
-rm Makefile
